update value objects too
