+++ config/set/php80.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 use Rector\Arguments\Rector\ClassMethod\ArgumentAdderRector;
 use Rector\Arguments\Rector\FuncCall\FunctionArgumentDefaultValueReplacerRector;
 use Rector\Arguments\ValueObject\ArgumentAdder;
-use Rector\Arguments\ValueObject\FuncCallArgumentDefaultValueReplacer;
+use Rector\Arguments\ValueObject\ReplaceFuncCallArgumentDefaultValue;
 use Rector\DeadCode\Rector\StaticCall\RemoveParentCallWithoutParentRector;
 use Rector\Php80\Rector\Catch_\RemoveUnusedVariableInCatchRector;
 use Rector\Php80\Rector\Class_\ClassPropertyAssignToConstructorPromotionRector;
@@ -97,15 +97,15 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $services->set(Rector\Arguments\Rector\FuncCall\FunctionArgumentDefaultValueReplacerRector::class)
         ->call('configure', [[
             FunctionArgumentDefaultValueReplacerRector::REPLACED_ARGUMENTS => ValueObjectInliner::inline([
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, 'gte', 'ge'),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, 'lte', 'le'),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, '', '!='),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, '!', '!='),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, 'g', 'gt'),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, 'l', 'lt'),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, 'gte', 'ge'),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, 'lte', 'le'),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, 'n', 'ne'),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, 'gte', 'ge'),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, 'lte', 'le'),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, '', '!='),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, '!', '!='),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, 'g', 'gt'),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, 'l', 'lt'),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, 'gte', 'ge'),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, 'lte', 'le'),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, 'n', 'ne'),
             ]),
         ]]);
 };
