+++ rules-tests/Arguments/Rector/ClassMethod/ReplaceArgumentDefaultValueRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Arguments\Rector\ClassMethod\ReplaceArgumentDefaultValueRector;
-use Rector\Arguments\ValueObject\ArgumentDefaultValueReplacer;
+use Rector\Arguments\ValueObject\ReplaceArgumentDefaultValue;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symplify\SymfonyPhpConfig\ValueObjectInliner;
 
@@ -13,35 +13,35 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         ->call('configure', [[
             ReplaceArgumentDefaultValueRector::REPLACED_ARGUMENTS => ValueObjectInliner::inline([
 
-                new ArgumentDefaultValueReplacer(
+                new ReplaceArgumentDefaultValue(
                     'Symfony\Component\DependencyInjection\Definition',
                     'setScope',
                     0,
                     'Symfony\Component\DependencyInjection\ContainerBuilder::SCOPE_PROTOTYPE',
                     false
                 ),
-                new ArgumentDefaultValueReplacer('Symfony\Component\Yaml\Yaml', 'parse', 1, [
+                new ReplaceArgumentDefaultValue('Symfony\Component\Yaml\Yaml', 'parse', 1, [
                     false,
                     false,
                     true,
                 ], 'Symfony\Component\Yaml\Yaml::PARSE_OBJECT_FOR_MAP'),
-                new ArgumentDefaultValueReplacer('Symfony\Component\Yaml\Yaml', 'parse', 1, [
+                new ReplaceArgumentDefaultValue('Symfony\Component\Yaml\Yaml', 'parse', 1, [
                     false,
                     true,
                 ], 'Symfony\Component\Yaml\Yaml::PARSE_OBJECT'),
-                new ArgumentDefaultValueReplacer('Symfony\Component\Yaml\Yaml', 'parse', 1, false, 0),
-                new ArgumentDefaultValueReplacer(
+                new ReplaceArgumentDefaultValue('Symfony\Component\Yaml\Yaml', 'parse', 1, false, 0),
+                new ReplaceArgumentDefaultValue(
                     'Symfony\Component\Yaml\Yaml',
                     'parse',
                     1,
                     true,
                     'Symfony\Component\Yaml\Yaml::PARSE_EXCEPTION_ON_INVALID_TYPE'
                 ),
-                new ArgumentDefaultValueReplacer('Symfony\Component\Yaml\Yaml', 'dump', 3, [
+                new ReplaceArgumentDefaultValue('Symfony\Component\Yaml\Yaml', 'dump', 3, [
                     false,
                     true,
                 ], 'Symfony\Component\Yaml\Yaml::DUMP_OBJECT'),
-                new ArgumentDefaultValueReplacer(
+                new ReplaceArgumentDefaultValue(
                     'Symfony\Component\Yaml\Yaml',
                     'dump',
                     3,
