+++ rules-tests/Arguments/Rector/FuncCall/FunctionArgumentDefaultValueReplacerRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Arguments\Rector\FuncCall\FunctionArgumentDefaultValueReplacerRector;
-use Rector\Arguments\ValueObject\FuncCallArgumentDefaultValueReplacer;
+use Rector\Arguments\ValueObject\ReplaceFuncCallArgumentDefaultValue;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symplify\SymfonyPhpConfig\ValueObjectInliner;
 
@@ -12,9 +12,9 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $services->set(FunctionArgumentDefaultValueReplacerRector::class)
         ->call('configure', [[
             FunctionArgumentDefaultValueReplacerRector::REPLACED_ARGUMENTS => ValueObjectInliner::inline([
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, 'lte', 'le'),
-                new FuncCallArgumentDefaultValueReplacer('version_compare', 2, '', '!='),
-                new FuncCallArgumentDefaultValueReplacer(
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, 'lte', 'le'),
+                new ReplaceFuncCallArgumentDefaultValue('version_compare', 2, '', '!='),
+                new ReplaceFuncCallArgumentDefaultValue(
                     'some_function',
                     0,
                     true,
